Replace AsyncTask with Coroutine
It's GlobalScope, which isn't good practice, but this is still better than AsyncTask

+++ app/src/main/java/org/jellyfin/androidtv/ui/playback/segments/SegmentSkipFragment.kt
@@ -46,7 +46,7 @@ class SegmentSkipFragment() : Fragment() {
 
 	private fun buttonClicked() {
 		lastSegment?.let { segment ->
-			playbackControllerContainer.playbackController?.seek((segment.endTime.millis).toLong())
+			playbackControllerContainer.playbackController?.seek((segment.endTime.millis))
 		}
 	}
